# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AdherenceOut": {
        "properties": {
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "has_schedules": {
            "title": "Has Schedules",
            "type": "boolean"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/AdherenceLineOut"
            },
            "title": "Lines",
            "type": "array"
          },
          "total_recorded": {
            "title": "Total Recorded",
            "type": "integer"
          }
        },
        "required": [
          "days",
          "has_schedules",
          "total_recorded",
          "lines"
        ],
        "title": "AdherenceOut",
        "type": "object"
      },
...
      "AsNeededIn": {
        "additionalProperties": false,
        "properties": {
          "taken_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken At"
          }
        },
        "title": "AsNeededIn",
        "type": "object"
      },
...
      "DoseActionIn": {
        "additionalProperties": false,
        "properties": {
          "minutes": {
            "anyOf": [
              {
                "enum": [
                  5,
                  10,
                  15,
                  30,
                  60
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "DoseActionIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Access token from /api/v1/auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Health Io API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/patients/{patient_id}/adherence": {
      "get": {
        "description": "Only recorded dose events; nothing is estimated.",
        "operationId": "adherence_api_v1_patients__patient_id__adherence_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdherenceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Adherence",
        "tags": [
          "medications"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/doses": {
      "get": {
        "description": "Doses for `days` calendar days from `day` (default: today in the patient's timezone).\nOpening the list also creates upcoming dose rows and records long-unanswered doses as\nmissed; the reminder engine (Phase 10) does the same on a schedule.",
        "operationId": "list_doses_api_v1_patients__patient_id__doses_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DoseOut"
                  },
                  "title": "Response List Doses Api V1 Patients  Patient Id  Doses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Doses",
        "tags": [
          "medications"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/doses/{dose_id}/{action}": {
      "post": {
        "operationId": "record_dose_api_v1_patients__patient_id__doses__dose_id___action__post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "dose_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Dose Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "enum": [
                "take",
                "skip",
                "snooze"
              ],
              "title": "Action",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoseActionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Dose",
        "tags": [
          "medications"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/medications": {
      "get": {
        "operationId": "list_medications_api_v1_patients__patient_id__medications_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MedicationOut"
                  },
                  "title": "Response List Medications Api V1 Patients  Patient Id  Medications Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Medications",
        "tags": [
          "medications"
        ]
      },
      "post": {
        "description": "Doctor: record a medicine the patient already takes (e.g. started elsewhere).\nNew treatment is written as a prescription instead.",
        "operationId": "record_existing_medication_api_v1_patients__patient_id__medications_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExistingMedicationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Existing Medication",
        "tags": [
          "medications"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/medications/self-reported": {
      "post": {
        "description": "A medicine the patient takes that no doctor prescribed here. Always labelled\npatient-reported.",
        "operationId": "add_self_reported_api_v1_patients__patient_id__medications_self_reported_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelfReportedIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Self Reported",
        "tags": [
          "medications"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/medications/{medication_id}/as-needed-dose": {
      "post": {
        "operationId": "log_as_needed_api_v1_patients__patient_id__medications__medication_id__as_needed_dose_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "medication_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Medication Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AsNeededIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Log As Needed",
        "tags": [
          "medications"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/medications/{medication_id}/confirm": {
      "post": {
        "description": "Start a prescribed medicine by choosing reminder times. The prescription itself\nis not changed.",
        "operationId": "confirm_medication_api_v1_patients__patient_id__medications__medication_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "medication_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Medication Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderTimesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Confirm Medication",
        "tags": [
          "medications"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/medications/{medication_id}/reminder-times": {
      "put": {
        "operationId": "change_reminder_times_api_v1_patients__patient_id__medications__medication_id__reminder_times_put",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "medication_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Medication Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderTimesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Reminder Times",
        "tags": [
          "medications"
        ]
      }
    },
...
  }
}